+++ src/index.ts
@@ -1,40 +1,20 @@
-import analyseVirtualFiles from './analyser/index.js';
-import fromFilesystem from './input/fromFilesystem.js';
-import fromRawContent from './input/fromRawContent.js';
-import retrieveData from './program/data/retrieveData.js';
-import * as T from './types/types.js';
+import analyseFs from './entry/analyseFs.js';
+import analyseRaw from './entry/analyseRaw.js';
 
-async function analyse(path?: string, opts?: T.Options): Promise<T.Results>;
-async function analyse(paths?: string[], opts?: T.Options): Promise<T.Results>;
-async function analyse(content?: Record<string, string>, opts?: T.Options): Promise<T.Results>;
-async function analyse(rawInput?: string | string[] | Record<string, string>, opts: T.Options = {}): Promise<T.Results> {
-	const inputs = {
-		path: typeof rawInput === 'string' ? rawInput : null,
-		paths: Array.isArray(rawInput) ? rawInput : null,
-		content: typeof rawInput === 'object' && !Array.isArray(rawInput) ? rawInput : null,
-	};
-	const inputPaths = inputs.paths ?? (inputs.path ? [inputs.path] : null);
-	const inputContent = inputs.content;
-
-	// Normalise input option arguments
-	opts = {
-		calculateLines: opts.calculateLines ?? true, // default to true if unset
-		checkIgnored: !opts.quick,
-		checkDetected: !opts.quick,
-		checkAttributes: !opts.quick,
-		checkHeuristics: !opts.quick,
-		checkShebang: !opts.quick,
-		checkModeline: !opts.quick,
-		...opts,
-	};
-
-	// Load data from github-linguist web repo
-	const { langData, heuristicsData, vendorPaths } = await retrieveData(opts.offline ?? false);
-
-	// Setup main variables
-	const files = inputContent ? fromRawContent(inputContent, vendorPaths) : await fromFilesystem(inputPaths ?? [], opts, vendorPaths);
-
-	return analyseVirtualFiles(files, langData, heuristicsData, opts);
-}
-
-export default analyse;
+// Entry point of the program
+export default {
+	/**
+	 * Analyse a list of folders and return language statistics
+	 * @param inputPaths - An array of file paths to analyse
+	 * @param options - Configuration options for the analysis
+	 * @returns A promise that resolves to the analysis results
+	 */
+	analyseFolders: analyseFs,
+	/**
+	 * Analyse raw content and return language statistics
+	 * @param inputContent - An object where keys are filenames and values are file contents
+	 * @param options - Configuration options for the analysis
+	 * @returns A promise that resolves to the analysis results
+	 */
+	analyseRawContent: analyseRaw,
+};