+++ test/unit.js
@@ -8,7 +8,7 @@ function desc(text) {
 }
 
 async function test([filename, fileContent = ''], [type, testVal]) {
-	const actual = await linguist({ [filename]: fileContent }, { childLanguages: true });
+	const actual = await linguist.analyseRawContent({ [filename]: fileContent }, { childLanguages: true });
 	const testContent = {
 		'files': actual.files.results[filename],
 		'size': actual.files.bytes,
@@ -59,7 +59,7 @@ async function unitTest() {
 	await test(['ecl.ecl', 'var:=val'], ['files', 'ECL']);
 	await test(['frege.fr', 'import package'], ['files', 'Frege']);
 	await test(['forth.fr', 'new-device 1'], ['files', 'Forth']);
-	await test(['raku','#!/usr/bin/env perl6\n module'], ['files', 'Raku']);
+	await test(['raku', '#!/usr/bin/env perl6\n module'], ['files', 'Raku']);
 	desc('vendored');
 	await test(['gradlew'], ['files', undefined]);
 	await test(['decl.d.ts'], ['files', undefined]);